Pass all tests through test/testing_suite_test.dart
All original test files have been altered in this way:

End with ".dart" instead of "_test.dart".

Reside in the "test/tests/" folder instead of the "test/" folder.

Export a function which sets up the client and runs all of the original tests (tests themselves have not been altered in any way).